physical address now matches output from readmap2 exactly.

+++ linux-4.1.15/include/linux/syscalls.h
@@ -894,4 +894,4 @@ asmlinkage long sys_execveat(int dfd, const char __user *filename,
 
 //asmlinkage long sys_get_proc_custom(int __user *proc_count, const struct CustomProcInfo __user *proc_array); 
 
-asmlinkage long sys_get_addr(int pid, unsigned long addr);
+asmlinkage long sys_get_addr(int pid, unsigned long long addr);
